+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.12.0] - 2025-08-01
+## [0.13.0] - 2025-08-05
+
+### Added
+
+- Added `listComponentVersions` tool to list all versions for a CircleCI component
+
+### Changed
+
+- Simplified `runRollbackPipeline` tool by moving part of its inner logic to `listComponentVersions`.
+
+## [0.12.2] - 2025-08-01
 
 ### Added